requirements.txt
openai>=1.0.0
SpeechRecognition>=3.10.0
pyttsx3>=2.90
python-dotenv>=1.0.0
PyAudio>=0.2.13
